+++ test/index.test.js
@@ -1,13 +1,17 @@
 import * as fs from 'fs'
 import * as zlib from 'zlib'
+import { promisify } from 'util'
 import test from 'ava'
 import rimraf from 'rimraf'
 import * as rollup from 'rollup'
 import gzip from '../dist/index'
 
+const writeFilePromise = promisify(fs.writeFile)
+const createDirPromise = promisify(fs.mkdir)
+
 function cleanup() {
     return new Promise(resolve => {
-        rimraf('test/__output', () => resolve())
+        rimraf('test/__output/**', () => resolve())
     })
 }
 
@@ -20,7 +24,7 @@ function fileNotPresent(t, path) {
     })
 }
 
-function compareFileWithGzip(t, path) {
+function compareFileWithGzip(t, path, extension) {
     return new Promise((resolve, reject) => {
         fs.readFile(path, (err, bundleContent) => {
             if (err) {
@@ -29,7 +33,7 @@ function compareFileWithGzip(t, path) {
                 return
             }
 
-            fs.readFile(path + '.gz', (err, gzipContent) => {
+            fs.readFile(path + (extension || '.gz'), (err, gzipContent) => {
                 if (err) {
                     t.fail('Gzip file not found!')
                     reject()
@@ -48,14 +52,26 @@ function compareFileWithGzip(t, path) {
     })
 }
 
-test.beforeEach(() => cleanup())
-//test.afterEach(() => cleanup())
+function writeFiles(files) {
+    return createDirPromise('test/__output').then(() =>
+        Promise.all(files.map(file => writeFilePromise(file[0], file[1]))),
+    )
+}
 
-test.serial('without options', t => {
-    return rollup
+function delay(ms) {
+    return new Promise((resolve, reject) =>
+        setTimeout(() => resolve(), ms || 2500),
+    )
+}
+
+test.beforeEach(() => cleanup())
+test.afterEach(() => cleanup())
+
+const sampleRollup = options =>
+    rollup
         .rollup({
             input: 'test/sample/index.js',
-            plugins: [gzip()],
+            plugins: [gzip(options)],
         })
         .then(bundle => {
             return bundle.write({
@@ -64,136 +80,13 @@ test.serial('without options', t => {
                 sourcemap: true,
             })
         })
-        .then(() => compareFileWithGzip(t, 'test/__output/bundle.js'))
-})
 
-test.skip.serial('with options', t => {
-    return rollup
-        .rollup({
-            input: 'test/sample/index.js',
-            plugins: [
-                // file that is above the size option => gets compressed
-                {
-                    name: 'test',
-                    onwrite: (options, bundle) => {
-                        return new Promise(resolve => {
-                            fs.writeFile(
-                                'test/__output/test1.txt',
-                                'This is a test',
-                                () => resolve(),
-                            )
-                        })
-                    },
-                },
-                // short file that is below the size option => not compressed
-                {
-                    name: 'test2',
-                    onwrite: (options, bundle) => {
-                        return new Promise(resolve => {
-                            fs.writeFile(
-                                'test/__output/test2.txt',
-                                'Short',
-                                () => resolve(),
-                            )
-                        })
-                    },
-                },
-                gzip({
-                    options: {
-                        level: 9,
-                    },
-                    additional: [
-                        'test/__output/test1.txt',
-                        'test/__output/test2.txt',
-                    ],
-                    minSize: 10,
-                }),
-            ],
-        })
-        .then(bundle => {
-            return bundle.write({
-                file: 'test/__output/bundle.js',
-                format: 'cjs',
-            })
-        })
-        .then(() => compareFileWithGzip(t, 'test/__output/bundle.js'))
-        .then(() => compareFileWithGzip(t, 'test/__output/test1.txt'))
-        .then(() => fileNotPresent(t, 'test/__output/test2.txt.gz'))
-})
-
-test.skip.serial('delayed writing in plugin', t => {
-    return rollup
-        .rollup({
-            input: 'test/sample/index.js',
-            plugins: [
-                {
-                    name: 'test',
-                    onwrite: (options, bundle) => {
-                        return new Promise(resolve => {
-                            setTimeout(() => {
-                                fs.writeFile(
-                                    'test/__output/test3.txt',
-                                    'This is a test',
-                                    () => resolve(),
-                                )
-                            }, 1000)
-                        })
-                    },
-                },
-                gzip({
-                    additional: ['test/__output/test3.txt'],
-                }),
-            ],
-        })
-        .then(bundle => {
-            return bundle.write({
-                file: 'test/__output/bundle.js',
-                format: 'cjs',
-            })
-        })
-        .then(() => compareFileWithGzip(t, 'test/__output/bundle.js'))
-        .then(() => compareFileWithGzip(t, 'test/__output/test3.txt'))
-})
-
-test.skip.serial('delayed writing in plugin without promise', t => {
-    return rollup
-        .rollup({
-            input: 'test/sample/index.js',
-            plugins: [
-                {
-                    name: 'test',
-                    onwrite: (options, bundle) => {
-                        setTimeout(() => {
-                            fs.writeFile(
-                                'test/__output/test3.txt',
-                                'This is a test',
-                                () => {},
-                            )
-                        }, 1000)
-                    },
-                },
-                gzip({
-                    additional: ['test/__output/test3.txt'],
-                    delay: 2000,
-                }),
-            ],
-        })
-        .then(bundle => {
-            return bundle.write({
-                file: 'test/__output/bundle.js',
-                format: 'cjs',
-            })
-        })
-        .then(() => compareFileWithGzip(t, 'test/__output/bundle.js'))
-        .then(() => compareFileWithGzip(t, 'test/__output/test3.txt'))
-})
-
-test.skip.serial('splitting', t => {
-    return rollup
+const sampleSplittingRollup = options =>
+    rollup
         .rollup({
             input: ['test/sample-splitting/a.js', 'test/sample-splitting/b.js'],
             experimentalCodeSplitting: true,
-            plugins: [gzip()],
+            plugins: [gzip(options)],
         })
         .then(bundle => {
             return bundle.write({
@@ -201,8 +94,106 @@ test.skip.serial('splitting', t => {
                 format: 'cjs',
             })
         })
-        .then(() => new Promise(resolve => setTimeout(() => resolve(), 1500)))
-        .then(() => compareFileWithGzip(t, 'test/__output/a.js'))
+
+test.serial('without options', t => {
+    return sampleRollup().then(() =>
+        compareFileWithGzip(t, 'test/__output/bundle.js'),
+    )
+})
+
+test.serial('with fileName string option', t => {
+    return sampleRollup({
+        fileName: '.gzzz',
+        algorithm: 'zopfli',
+    }).then(() => compareFileWithGzip(t, 'test/__output/bundle.js', '.gzzz'))
+})
+
+test.serial('with fileName function option', t => {
+    return sampleRollup({
+        fileName: name => name + '.gzzz',
+    }).then(() => compareFileWithGzip(t, 'test/__output/bundle.js', '.gzzz'))
+})
+
+test.serial('with customCompression option', t => {
+    return sampleRollup({
+        customCompression: content => content.toString() + 'XXX',
+    }).then(
+        () =>
+            new Promise((resolve, reject) => {
+                fs.readFile('test/__output/bundle.js', (err, bundleContent) => {
+                    if (err) {
+                        t.fail('Bundle not found!')
+                        reject()
+                        return
+                    }
+
+                    fs.readFile(
+                        'test/__output/bundle.js.gz',
+                        (err, compressedContent) => {
+                            if (err) {
+                                t.fail('Gzip file not found!')
+                                reject()
+                                return
+                            }
+
+                            t.deepEqual(
+                                compressedContent.toString(),
+                                bundleContent.toString() + 'XXX',
+                            )
+                            resolve()
+                        },
+                    )
+                })
+            }),
+    )
+})
+
+test.serial('with additionalFiles option', t => {
+    return writeFiles([
+        ['test/__output/test1.txt', 'This is a test'],
+        ['test/__output/test2.txt', 'This too'],
+    ])
+        .then(() =>
+            sampleRollup({
+                gzipOptions: {
+                    level: 9,
+                },
+                additionalFiles: [
+                    'test/__output/test1.txt',
+                    'test/__output/test2.txt',
+                ],
+                additionalFilesDelay: 2000,
+            }),
+        )
+        .then(() => compareFileWithGzip(t, 'test/__output/bundle.js'))
+        .then(() => delay())
+        .then(() => compareFileWithGzip(t, 'test/__output/test1.txt'))
+        .then(() => compareFileWithGzip(t, 'test/__output/test2.txt'))
+})
+
+test.serial('splitting with regex filter option', t => {
+    return sampleSplittingRollup({
+        filter: /(b|c).js$/,
+    })
+        .then(() => fileNotPresent(t, 'test/__output/a.js.gz'))
         .then(() => compareFileWithGzip(t, 'test/__output/b.js'))
         .then(() => compareFileWithGzip(t, 'test/__output/c.js'))
+})
+
+test.serial('splitting with function filter option', t => {
+    return sampleSplittingRollup({
+        filter: fileName => fileName[0] !== 'a',
+    })
+        .then(() => fileNotPresent(t, 'test/__output/a.js.gz'))
+        .then(() => compareFileWithGzip(t, 'test/__output/b.js'))
+        .then(() => compareFileWithGzip(t, 'test/__output/c.js'))
+})
+
+test.serial('splitting with minSize option', t => {
+    return sampleSplittingRollup({
+        minSize: 80,
+    })
+        .then(() => fileNotPresent(t, 'test/__output/c.js.gz'))
+        .then(() => compareFileWithGzip(t, 'test/__output/a.js'))
+        .then(() => compareFileWithGzip(t, 'test/__output/b.js'))
 })